remove comfy_table from rustic_core
