+++ crates/rustic_core/src/repofile/snapshotfile.rs
@@ -18,14 +18,6 @@ use serde::{Deserialize, Serialize};
 
 use serde_with::{serde_as, DeserializeFromStr, DisplayFromStr};
 
-#[cfg(feature = "cli")]
-use crate::helpers::{
-    bytes_size_to_string,
-    table_output::{bold_cell, table, Cell},
-};
-#[cfg(feature = "cli")]
-use humantime::format_duration;
-
 use crate::{
     backend::{decrypt::DecryptReadBackend, FileType},
     error::SnapshotFileErrorKind,
@@ -473,97 +465,6 @@ impl PartialOrd for SnapshotFile {
 impl Ord for SnapshotFile {
     fn cmp(&self, other: &Self) -> Ordering {
         self.time.cmp(&other.time)
-    }
-}
-
-#[cfg(feature = "cli")]
-impl Display for SnapshotFile {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut table = table();
-
-        let mut add_entry = |title: &str, value: String| {
-            _ = table.add_row([bold_cell(title), Cell::new(value)]);
-        };
-
-        add_entry("Snapshot", self.id.to_hex().to_string());
-        // note that if original was not set, it is set to self.id by the load process
-        if self.original != Some(self.id) {
-            add_entry("Original ID", self.original.unwrap().to_hex().to_string());
-        }
-        add_entry("Time", self.time.format("%Y-%m-%d %H:%M:%S").to_string());
-        add_entry("Generated by", self.program_version.clone());
-        add_entry("Host", self.hostname.clone());
-        add_entry("Label", self.label.clone());
-        add_entry("Tags", self.tags.formatln());
-        let delete = match self.delete {
-            DeleteOption::NotSet => "not set".to_string(),
-            DeleteOption::Never => "never".to_string(),
-            DeleteOption::After(t) => format!("after {}", t.format("%Y-%m-%d %H:%M:%S")),
-        };
-        add_entry("Delete", delete);
-        add_entry("Paths", self.paths.formatln());
-        let parent = self.parent.map_or_else(
-            || "no parent snapshot".to_string(),
-            |p| p.to_hex().to_string(),
-        );
-        add_entry("Parent", parent);
-        if let Some(ref summary) = self.summary {
-            add_entry("", String::new());
-            add_entry("Command", summary.command.clone());
-
-            let source = format!(
-                "files: {} / dirs: {} / size: {}",
-                summary.total_files_processed,
-                summary.total_dirs_processed,
-                bytes_size_to_string(summary.total_bytes_processed)
-            );
-            add_entry("Source", source);
-            add_entry("", String::new());
-
-            let files = format!(
-                "new: {:>10} / changed: {:>10} / unchanged: {:>10}",
-                summary.files_new, summary.files_changed, summary.files_unmodified,
-            );
-            add_entry("Files", files);
-
-            let trees = format!(
-                "new: {:>10} / changed: {:>10} / unchanged: {:>10}",
-                summary.dirs_new, summary.dirs_changed, summary.dirs_unmodified,
-            );
-            add_entry("Dirs", trees);
-            add_entry("", String::new());
-
-            let written = format!(
-                "data:  {:>10} blobs / raw: {:>10} / packed: {:>10}\n\
-            tree:  {:>10} blobs / raw: {:>10} / packed: {:>10}\n\
-            total: {:>10} blobs / raw: {:>10} / packed: {:>10}",
-                summary.data_blobs,
-                bytes_size_to_string(summary.data_added_files),
-                bytes_size_to_string(summary.data_added_files_packed),
-                summary.tree_blobs,
-                bytes_size_to_string(summary.data_added_trees),
-                bytes_size_to_string(summary.data_added_trees_packed),
-                summary.tree_blobs + summary.data_blobs,
-                bytes_size_to_string(summary.data_added),
-                bytes_size_to_string(summary.data_added_packed),
-            );
-            add_entry("Added to repo", written);
-
-            let duration = format!(
-                "backup start: {} / backup end: {} / backup duration: {}\n\
-            total duration: {}",
-                summary.backup_start.format("%Y-%m-%d %H:%M:%S"),
-                summary.backup_end.format("%Y-%m-%d %H:%M:%S"),
-                format_duration(std::time::Duration::from_secs_f64(summary.backup_duration)),
-                format_duration(std::time::Duration::from_secs_f64(summary.total_duration))
-            );
-            add_entry("Duration", duration);
-        }
-        if let Some(ref description) = self.description {
-            add_entry("Description", description.clone());
-        }
-
-        write!(f, "{table}")
     }
 }
