+++ src/commands/prune.rs
@@ -5,6 +5,7 @@
 use crate::{
     commands::{get_repository, open_repository},
     config::progress_options::ProgressOptions,
+    helpers::bytes_size_to_string,
     status_err, Application, RUSTIC_APP,
 };
 use abscissa_core::{Command, Runnable, Shutdown};
@@ -28,10 +29,10 @@ use rayon::prelude::{IntoParallelIterator, ParallelIterator};
 use crate::helpers::warm_up_wait;
 
 use rustic_core::{
-    bytes_size_to_string, BlobType, BlobTypeMap, DecryptReadBackend, DecryptWriteBackend, FileType,
-    HeaderEntry, Id, IndexBackend, IndexBlob, IndexCollector, IndexFile, IndexPack, IndexType,
-    IndexedBackend, Indexer, Initialize, NodeType, OpenRepository, PackSizer, Progress,
-    ReadBackend, ReadIndex, Repacker, SnapshotFile, Sum, TreeStreamerOnce,
+    BlobType, BlobTypeMap, DecryptReadBackend, DecryptWriteBackend, FileType, HeaderEntry, Id,
+    IndexBackend, IndexBlob, IndexCollector, IndexFile, IndexPack, IndexType, IndexedBackend,
+    Indexer, Initialize, NodeType, OpenRepository, PackSizer, Progress, ReadBackend, ReadIndex,
+    Repacker, SnapshotFile, Sum, TreeStreamerOnce,
 };
 
 pub(super) mod constants {
