+++ src/helpers.rs
@@ -2,6 +2,10 @@ use std::{collections::BTreeSet, process::Command};
 
 use abscissa_core::Application;
 use anyhow::Result;
+use bytesize::ByteSize;
+use comfy_table::{
+    presets::ASCII_MARKDOWN, Attribute, Cell, CellAlignment, ContentArrangement, Table,
+};
 
 use log::{debug, info, trace, warn};
 use rayon::{
@@ -10,9 +14,9 @@ use rayon::{
 };
 
 use rustic_core::{
-    parse_command, table_with_titles, BlobType, DecryptWriteBackend, FileType, Id, IndexBackend,
-    IndexedBackend, Indexer, NodeType, OpenRepository, Packer, Progress, ReadBackend, ReadIndex,
-    SnapshotFile, TreeStreamerOnce,
+    parse_command, BlobType, DecryptWriteBackend, FileType, Id, IndexBackend, IndexedBackend,
+    Indexer, NodeType, OpenRepository, Packer, Progress, ReadBackend, ReadIndex, RusticResult,
+    SnapshotFile, TreeStreamerOnce, ALL_FILE_TYPES,
 };
 
 use crate::{application::RUSTIC_APP, config::progress_options::ProgressOptions};
@@ -236,4 +240,72 @@ where
     println!("{table}");
 
     Ok(snaps)
+}
+
+/// Helpers for table output
+
+pub fn bold_cell<T: ToString>(s: T) -> Cell {
+    Cell::new(s).add_attribute(Attribute::Bold)
+}
+
+#[must_use]
+pub fn table() -> Table {
+    let mut table = Table::new();
+    _ = table
+        .load_preset(ASCII_MARKDOWN)
+        .set_content_arrangement(ContentArrangement::Dynamic);
+    table
+}
+
+pub fn table_with_titles<I: IntoIterator<Item = T>, T: ToString>(titles: I) -> Table {
+    let mut table = table();
+    _ = table.set_header(titles.into_iter().map(bold_cell));
+    table
+}
+
+pub fn table_right_from<I: IntoIterator<Item = T>, T: ToString>(start: usize, titles: I) -> Table {
+    let mut table = table_with_titles(titles);
+    // set alignment of all rows except first start row
+    table
+        .column_iter_mut()
+        .skip(start)
+        .for_each(|c| c.set_cell_alignment(CellAlignment::Right));
+
+    table
+}
+
+pub fn print_file_info(text: &str, be: &impl ReadBackend) -> RusticResult<()> {
+    info!("scanning files...");
+
+    let mut table = table_right_from(1, ["File type", "Count", "Total Size"]);
+    let mut total_count = 0;
+    let mut total_size = 0;
+    for tpe in ALL_FILE_TYPES {
+        let list = be.list_with_size(tpe)?;
+        let count = list.len();
+        let size = list.iter().map(|f| u64::from(f.1)).sum();
+        _ = table.add_row([
+            format!("{tpe:?}"),
+            count.to_string(),
+            bytes_size_to_string(size),
+        ]);
+        total_count += count;
+        total_size += size;
+    }
+    println!("{text}");
+    _ = table.add_row([
+        "Total".to_string(),
+        total_count.to_string(),
+        bytes_size_to_string(total_size),
+    ]);
+
+    println!();
+    println!("{table}");
+    println!();
+    Ok(())
+}
+
+#[must_use]
+pub fn bytes_size_to_string(b: u64) -> String {
+    ByteSize(b).to_string_as(true)
 }
